Kadir

+++ database/createTables.sql
@@ -75,7 +75,7 @@ CREATE TABLE IF NOT EXISTS Course(
 -- password,name,surname,email,department_id should be taken from Users table using username key.
 CREATE TABLE IF NOT EXISTS Students(
     username CHAR(100),
-    student_id INTEGER,
+    student_id BIGINT,
     -- added_courses will be used as JSON_ARRAY
     added_courses JSON,
     completed_credits INT DEFAULT 0,
@@ -89,7 +89,7 @@ CREATE TABLE IF NOT EXISTS Students(
 -- Create Grades table if not exist. primary keys are student_id and course_id,
 -- which are also referanced from other tables.
 CREATE TABLE IF NOT EXISTS Grades(
-    student_id INTEGER,
+    student_id BIGINT,
     course_id CHAR(100),
     grade FLOAT,
     PRIMARY KEY (course_id, student_id),
@@ -97,6 +97,8 @@ CREATE TABLE IF NOT EXISTS Grades(
     FOREIGN KEY (course_id) REFERENCES Course(course_id) 
 );
 
+-- Create a procedure "filter" if it does not exist.
+-- Take the 4 parameters, then filter the data based on those.
 DROP PROCEDURE IF EXISTS filter;
 CREATE PROCEDURE filter( IN department_id CHAR(100), 
                         IN campus CHAR(100), 
@@ -126,6 +128,8 @@ END;
 
 -- Delete triger if exists to prevent error.
 DROP TRIGGER IF EXISTS chk_quota;
+-- This trigger checks the relation between quota and classrom capacity.
+-- It checks and gives error if there is conflict between them.
 CREATE TRIGGER chk_quota BEFORE INSERT ON Course FOR EACH ROW
 BEGIN
     SELECT s.classroom_capacity INTO @capacity FROM Classroom s
@@ -139,6 +143,8 @@ END;
 
 -- Delete triger if exists to prevent error.
 DROP TRIGGER IF EXISTS chk_student;
+-- This trigger sets the gpa with the calculation given whenever
+-- there is an insertion to the talble grades.
 CREATE TRIGGER chk_student AFTER INSERT ON Grades FOR EACH ROW
 BEGIN
     SELECT SUM(c.credits), SUM(c.credits*g.grade)/SUM(c.credits) INTO @credits, @gpa
